requirements.txt
cassandra-driver
database-connect
scikit-learn
pandas
numpy
kneed
matplotlib
xgboost
flask
flask_cors
flask_monitoringdashboard
gunicorn
-e .